Various fixes and changes
