+++ Dyalect/Parser/Dyalect.atg
@@ -73,8 +73,8 @@ PRODUCTIONS
             | charToken (. val = ParseChar(); .)
         ).
 
-    StandardOperators = "+" | "-" | "*" | "/" | "%" | "|||" | "&&&" | "!" | "#" 
-        | "==" | "!=" | ">" | "<" | ">=" | "<=" | "^^^" | "<<<" | ">>>" | "~~~".
+    StandardOperators = "+" | "-" | "*" | "/" | "%" | "|" | "&" | "!" | "#" 
+        | "==" | "!=" | ">" | "<" | ">=" | "<=" | "^" | "<<" | ">>" | "~".
 
     FunctionName = identToken | StandardOperators.
 
@@ -362,11 +362,11 @@ PRODUCTIONS
                 | "*="  (. op = BinaryOperator.Mul; .)
                 | "/="  (. op = BinaryOperator.Div; .)
                 | "%="  (. op = BinaryOperator.Rem; .)
-                | "&&&="  (. op = BinaryOperator.And; .)
-                | "|||="  (. op = BinaryOperator.Or; .)
-                | "^^^="  (. op = BinaryOperator.Xor; .)
-                | "<<<=" (. op = BinaryOperator.ShiftLeft; .)
-                | ">>>=" (. op = BinaryOperator.ShiftRight; .)
+                | "&="  (. op = BinaryOperator.And; .)
+                | "|="  (. op = BinaryOperator.Or; .)
+                | "^="  (. op = BinaryOperator.Xor; .)
+                | "<<=" (. op = BinaryOperator.ShiftLeft; .)
+                | ">>=" (. op = BinaryOperator.ShiftRight; .)
             )
             Expr<out node>
             (.
@@ -383,10 +383,10 @@ PRODUCTIONS
             (. node = new DBinaryOperation(node, exp, BinaryOperator.Or, ot); .)
         }
         [
-            "if"
+            "?"
             (. var @if = new DIf(t) { True = node }; .)
             Or<out node> (. @if.Condition = node; .)
-            "else" Or<out node>
+            ":" Or<out node>
             (. @if.False = node; node = @if; .)
         ].
         
@@ -424,8 +424,8 @@ PRODUCTIONS
                 var ot = default(Token);
             .)
             ( 
-                  "<<<" (. ot = t; op = BinaryOperator.ShiftLeft; .)
-                | ">>>" (. ot = t; op = BinaryOperator.ShiftRight; .)
+                  "<<" (. ot = t; op = BinaryOperator.ShiftLeft; .)
+                | ">>" (. ot = t; op = BinaryOperator.ShiftRight; .)
             )
             BitOr<out DNode exp>
             (. node = new DBinaryOperation(node, exp, op, ot); .)
@@ -434,21 +434,21 @@ PRODUCTIONS
     BitOr<out DNode node> = 
         Xor<out node>
         {
-            "|||" (. var ot = t; .) Xor<out DNode exp>
+            "|" (. var ot = t; .) Xor<out DNode exp>
             (. node = new DBinaryOperation(node, exp, BinaryOperator.BitwiseOr, ot); .)
         }.
         
     Xor<out DNode node> = 
         BitAnd<out node>
         {
-            "^^^" (. var ot = t; .) BitAnd<out DNode exp>
+            "^" (. var ot = t; .) BitAnd<out DNode exp>
             (. node = new DBinaryOperation(node, exp, BinaryOperator.Xor, ot); .)
         }.
         
     BitAnd<out DNode node> = 
         Add<out node>
         {
-            "&&&" (. var ot = t; .) Add<out DNode exp>
+            "&" (. var ot = t; .) Add<out DNode exp>
             (. node = new DBinaryOperation(node, exp, BinaryOperator.BitwiseAnd, ot); .)
         }.
