+++ Dyalect/Runtime/Types/DyIterator.cs
@@ -2,11 +2,12 @@
 
 namespace Dyalect.Runtime.Types
 {
-    public sealed class DyIterator : DyFunction
+    internal sealed class DyIterator : DyForeignFunction
     {
+        internal const string Name = "iterator";
         private readonly IEnumerator<DyObject> enumerator;
 
-        public DyIterator(IEnumerator<DyObject> enumerator) : base(0, ExternId, 0, null, null, StandardType.Iterator)
+        public DyIterator(IEnumerator<DyObject> enumerator) : base(Name, 0, StandardType.Iterator)
         {
             this.enumerator = enumerator;
         }
@@ -18,7 +19,7 @@ namespace Dyalect.Runtime.Types
             return new DyNativeIterator(unitId, handle, vm, vars);
         }
 
-        protected override string GetFunctionName() => "iterator";
+        protected override string GetFunctionName() => Name;
 
         public override DyObject Call(ExecutionContext ctx, params DyObject[] args)
         {
@@ -27,17 +28,17 @@ namespace Dyalect.Runtime.Types
             return DyNil.Terminator;
         }
 
-        internal override DyFunction Clone(DyObject arg) =>
-            new DyIterator(enumerator) { Self = arg, Flags = Flags };
+        internal override DyFunction Clone(DyObject arg) => new DyIterator(enumerator) { Self = arg };
     }
-    internal sealed class DyNativeIterator : DyFunction
+
+    internal sealed class DyNativeIterator : DyNativeFunction
     {
         public DyNativeIterator(int unitId, int funcId, DyMachine vm, FastList<DyObject[]> captures) : base(unitId, funcId, 0, vm, captures, StandardType.Iterator)
         {
 
         }
 
-        protected override string GetFunctionName() => "iterator";
+        protected override string GetFunctionName() => DyIterator.Name;
 
         internal override DyFunction Clone(DyObject arg) =>
             new DyNativeIterator(UnitId, FunctionId, Machine, Captures) { Self = arg };
@@ -54,6 +55,6 @@ namespace Dyalect.Runtime.Types
 
         public override string TypeName => StandardType.BoolName;
 
-        protected override DyString ToStringOp(DyObject arg, ExecutionContext ctx) => "iterator()";
+        protected override DyString ToStringOp(DyObject arg, ExecutionContext ctx) => $"{DyIterator.Name}()";
     }
 }
